+++ tests/python/test_loops.py
@@ -44,8 +44,8 @@ def test_numpy_loops():
         y[i] = i - 300
 
     import numpy as np
-    begin = (np.ones(1) * (N // 2 + 3)).astype(np.int32)
-    end = (np.ones(1) * N).astype(np.int32)
+    begin = (np.ones(1) * (N // 2 + 3)).astype(np.int32).reshape(())
+    end = (np.ones(1) * N).astype(np.int32).reshape(())
 
     @ti.kernel
     def func():
