+++ tests/python/test_type_check.py
@@ -1,5 +1,6 @@
 import pytest
 
+import numpy as np
 import taichi as ti
 
 
@@ -53,3 +54,27 @@ def test_subscript():
 
     with pytest.raises(ti.TaichiTypeError, match="indices must be integers"):
         any_array(a)
+
+
+@ti.test()
+def test_0d_ndarray():
+    @ti.kernel
+    def foo() -> ti.i32:
+        a = np.array(3, dtype=np.int32)
+        return a
+
+    assert foo() == 3
+
+
+@ti.test()
+def test_non_0d_ndarray():
+    @ti.kernel
+    def foo():
+        a = np.array([1])
+
+    with pytest.raises(
+            ti.TaichiTypeError,
+            match=
+            "Only 0-dimensional numpy array can be used to initialize a scalar expression"
+    ):
+        foo()
